Add requirements for time freeze
We need a new map, so map002 is the designated "event" map. This map defines the event which is spawned on time freeze usage. Likewise, system.json is updated to reflect this new map. In plugin.js, I removed MadeWithMV and added the event spawning plugin and the time freeze plugin

un-git-ignore