Add logic for freezing buttons and doors
More complexity than expected. Simple API, in a button you will put <frs_isButton:selfSwitchOn,selfSwitchOff,selfSwitchOnCondition> where each is a single letter corresponding to the relevant self switch. The condition switch is the self switch which determines if a button is pressed or not.
Likewise for doors, <frs_isDoor:selfSwitchOpen,selfSwitchClosed,globalSwitchOpenCondition> where the first two are similar as buttons, but the globalSwitchOpenCondition is a global switch value as a number