Restart rendering if we failed to render in so many frames
Attempts to fix a bug that Kristen ran into. It's unreproducable, but the only reason it could of happened was because of failing to render to texture. If the problem is in PIXI we can't do anything, but if it's something to do with not calling frs.postProcess this will fix it